consistent device ID and water time
